chore(rust): fix compilation errors
