--- rust/op-reth/crates/tests/artifacts/.gitignore
@@ -1,2 +1,0 @@
-forge-artifacts
-src
