--- rust/op-reth/crates/tests/artifacts/compressed/README.md
@@ -1,2 +1,0 @@
-Artifacts in this directory will be embedded inside the `op-deployer` binary. The directory can be populated by running
-`make unzip-contract-artifacts`.
