+++ rust/op-reth/tests/proofs/contracts/foundry.toml
@@ -1,0 +1,3 @@
+[profile.default]
+src = "src"
+out = "artifacts"
